Add files via upload
Create the Retail Store Application. A retail store wants to know its top customers 
of each week, that is, the first topN customers with the largest sales, where topN 
is a value that the user of the program supplies. 
Implement a function 
 
 def nameOfBestCustomers(sales, customers, topN) 
 
that returns the names of the topN customers with the largest sales. 
The application prompts the user to enter a value for topN. Then it asks for all 
prices and names, calls the function that you implemented, and displays the 
results. If there were fewer than topN customers entered by the user, the 
application should output all the customers entered. Use character ‘x’ as a 
sentinel. The program output should be formatted as shown in the Sample Run.